Separate PluginTask and HostTask to another Class to make code clear
